+++ public/index.html
@@ -12,7 +12,7 @@
   <script src="https://www.amcharts.com/lib/4/core.js"></script>
   <script src="https://www.amcharts.com/lib/4/charts.js"></script>
   <script src="https://www.amcharts.com/lib/4/themes/animated.js"></script>
-  <script src="./javascripts/chart.js"></script>
+  <script type= 'module' src="./javascripts/main.js"></script>
